+++ packages/formik/src/useEnhancedFormikContext.ts
@@ -1,18 +1,24 @@
-import { useFormikContext, FormikContextType } from "formik";
-import { getApplyErrorsToFields, ErrorHelpers } from "./errors";
-import { getSetFormSuccess, getSetFormError, StatusHelpers } from "./status";
-
-type EnhancedFormikContextType<FormikValues> = FormikContextType<FormikValues> &
-  StatusHelpers &
-  ErrorHelpers;
+import { useFormikContext } from "formik";
+import { getApplyErrorsToFields } from "./errors";
+import {
+  getEnhancedSetStatus,
+  getSetFormSuccess,
+  getSetFormError,
+  getSetSentryEventId,
+} from "./status";
+import { EnhancedFormikContextType } from "./types";
 
 export function useEnhancedFormikContext<FormikValues>(): EnhancedFormikContextType<FormikValues> {
   const context = useFormikContext<FormikValues>();
 
+  const setStatus = getEnhancedSetStatus(context.setStatus, context.status);
+
   return {
     ...context,
     applyErrorsToFields: getApplyErrorsToFields(context.setErrors),
-    setFormSuccess: getSetFormSuccess(context.setStatus),
-    setFormError: getSetFormError(context.setStatus),
+    setStatus,
+    setFormSuccess: getSetFormSuccess(setStatus),
+    setFormError: getSetFormError(setStatus),
+    setSentryEventId: getSetSentryEventId(setStatus),
   };
 }